# backend/openapi.yml
openapi: 3.1.0
info:
  title: Fahrtenbuch API
  version: 1.0.0-dev
  description: |
    REST-API für das selbst gehostete Fahrtenbuch. Die API umfasst
    Authentifizierung, MFA, Passkeys, App-Pairing, Geräteverwaltung,
    Benutzer, Fahrzeuge, Fahrten, Tracking, Statistiken, Import/Export,
    Geocoding sowie zentrale Kartendienst-Konfiguration.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
    description: Relativ zum Fahrtenbuch-Server
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Eigenes Profil abrufen
      responses:
        '200':
          description: Profil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    put:
      tags: [Users]
      summary: Eigenes Profil aktualisieren
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Aktualisiertes Profil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /users/me/password:
    put:
      tags: [Users]
      summary: Eigenes Passwort ändern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8 }
      responses:
        '200': { description: Passwort geändert }
...
  /users/me/devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: Gerät und zugehörige Sitzungen entfernen
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Gerät entfernt }
...
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Benutzer abrufen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Benutzer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    put:
      tags: [Admin]
      summary: Benutzer aktualisieren
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Benutzer aktualisiert }
    delete:
      tags: [Admin]
      summary: Benutzer löschen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Benutzer gelöscht }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        loginName: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        role: { type: string, enum: [user, admin] }
        locale: { type: string, default: de }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserUpdate:
      type: object
      properties:
        loginName: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        locale: { type: string }
...